+++ crates/nostr/src/nips/mod.rs
@@ -31,6 +31,7 @@ pub mod nip34;
 pub mod nip35;
 pub mod nip38;
 pub mod nip39;
+pub mod nip42;
 #[cfg(feature = "nip44")]
 pub mod nip44;
 #[cfg(all(feature = "std", feature = "nip46"))]